refactor(alpha-test): modularize runtime/execution/provider architecture
- split spec_runtime bridge/cache/runtime-policy internals into dedicated modules
- split spec_execution tool search, security scan, approval policy and bridge policy hashing paths
- centralize blocked report assembly in execute_spec
- split provider model selection/payload adaptation/error policy paths
- add architecture boundary and stress test helper scripts plus Taskfile/docs updates
