+++ crates/app/src/provider/mod.rs
@@ -5,36 +5,55 @@ use tokio::time::sleep;
 
 use crate::CliResult;
 
-use super::config::{LoongClawConfig, ProviderConfig, ProviderKind, ReasoningEffort};
+use super::config::LoongClawConfig;
 #[cfg(feature = "memory-sqlite")]
 use super::memory;
 
+mod error_policy;
+mod model_selection;
+mod payload_adaptation;
 mod policy;
 mod shape;
 mod transport;
 
+use error_policy::{
+    provider_request_failed_for_all_models, validate_provider_configuration,
+    validate_provider_feature_gate,
+};
+use model_selection::{fetch_available_models_with_policy, resolve_request_models};
+use payload_adaptation::{
+    adapt_payload_mode_for_error, build_completion_request_body, build_turn_request_body,
+    parse_provider_api_error, should_disable_tool_schema_for_error, should_try_next_model_on_error,
+    CompletionPayloadMode,
+};
+
 pub use shape::extract_provider_turn;
 
-pub fn build_messages_for_session(
+pub fn build_system_message(
     config: &LoongClawConfig,
-    session_id: &str,
     include_system_prompt: bool,
-) -> CliResult<Vec<Value>> {
-    let mut messages = Vec::new();
-    if include_system_prompt {
-        let system = config.cli.system_prompt.trim();
-        let snapshot = super::tools::capability_snapshot();
-        let content = if system.is_empty() {
-            snapshot
-        } else {
-            format!("{system}\n\n{snapshot}")
-        };
-        messages.push(json!({
-            "role": "system",
-            "content": content,
-        }));
+) -> Option<Value> {
+    if !include_system_prompt {
+        return None;
     }
 
+    let system = config.cli.system_prompt.trim();
+    let snapshot = super::tools::capability_snapshot();
+    let content = if system.is_empty() {
+        snapshot
+    } else {
+        format!("{system}\n\n{snapshot}")
+    };
+    Some(json!({
+        "role": "system",
+        "content": content,
+    }))
+}
+
+pub fn load_memory_window_messages(
+    config: &LoongClawConfig,
+    session_id: &str,
+) -> CliResult<Vec<Value>> {
     #[cfg(feature = "memory-sqlite")]
     {
         let mem_config = super::memory::runtime_config::MemoryRuntimeConfig {
@@ -42,17 +61,32 @@ pub fn build_messages_for_session(
         };
         let turns = memory::window_direct(session_id, config.memory.sliding_window, &mem_config)
             .map_err(|error| format!("load memory window failed: {error}"))?;
+        let mut messages = Vec::with_capacity(turns.len());
         for turn in turns {
             messages.push(json!({
                 "role": turn.role,
                 "content": turn.content,
             }));
         }
+        Ok(messages)
     }
     #[cfg(not(feature = "memory-sqlite"))]
     {
-        let _ = session_id;
+        let _ = (config, session_id);
+        Ok(Vec::new())
     }
+}
+
+pub fn build_messages_for_session(
+    config: &LoongClawConfig,
+    session_id: &str,
+    include_system_prompt: bool,
+) -> CliResult<Vec<Value>> {
+    let mut messages = Vec::new();
+    if let Some(system_message) = build_system_message(config, include_system_prompt) {
+        messages.push(system_message);
+    }
+    messages.extend(load_memory_window_messages(config, session_id)?);
     Ok(messages)
 }
 
@@ -65,22 +99,17 @@ pub async fn request_completion(config: &LoongClawConfig, messages: &[Value]) ->
     let request_policy = policy::ProviderRequestPolicy::from_config(&config.provider);
     let client = build_http_client(&request_policy)?;
     let model_candidates = resolve_request_models(config, &headers, &request_policy).await?;
-    let auto_model_mode = config.provider.model_selection_requires_fetch();
+    let request_context = ProviderRequestContext {
+        endpoint: &endpoint,
+        headers: &headers,
+        request_policy: &request_policy,
+        client: &client,
+        auto_model_mode: config.provider.model_selection_requires_fetch(),
+    };
 
     let mut last_error = None;
     for (index, model) in model_candidates.iter().enumerate() {
-        match request_completion_with_model(
-            config,
-            messages,
-            model,
-            auto_model_mode,
-            &endpoint,
-            &headers,
-            &request_policy,
-            &client,
-        )
-        .await
-        {
+        match request_completion_with_model(config, messages, model, &request_context).await {
             Ok(content) => return Ok(content),
             Err(model_error) => {
                 if model_error.try_next_model && index + 1 < model_candidates.len() {
@@ -91,8 +120,7 @@ pub async fn request_completion(config: &LoongClawConfig, messages: &[Value]) ->
             }
         }
     }
-    Err(last_error
-        .unwrap_or_else(|| "provider request failed for every model candidate".to_owned()))
+    Err(provider_request_failed_for_all_models(last_error))
 }
 
 pub async fn request_turn(
@@ -107,22 +135,17 @@ pub async fn request_turn(
     let request_policy = policy::ProviderRequestPolicy::from_config(&config.provider);
     let client = build_http_client(&request_policy)?;
     let model_candidates = resolve_request_models(config, &headers, &request_policy).await?;
-    let auto_model_mode = config.provider.model_selection_requires_fetch();
+    let request_context = ProviderRequestContext {
+        endpoint: &endpoint,
+        headers: &headers,
+        request_policy: &request_policy,
+        client: &client,
+        auto_model_mode: config.provider.model_selection_requires_fetch(),
+    };
 
     let mut last_error = None;
     for (index, model) in model_candidates.iter().enumerate() {
-        match request_turn_with_model(
-            config,
-            messages,
-            model,
-            auto_model_mode,
-            &endpoint,
-            &headers,
-            &request_policy,
-            &client,
-        )
-        .await
-        {
+        match request_turn_with_model(config, messages, model, &request_context).await {
             Ok(turn) => return Ok(turn),
             Err(model_error) => {
                 if model_error.try_next_model && index + 1 < model_candidates.len() {
@@ -133,8 +156,7 @@ pub async fn request_turn(
             }
         }
     }
-    Err(last_error
-        .unwrap_or_else(|| "provider request failed for every model candidate".to_owned()))
+    Err(provider_request_failed_for_all_models(last_error))
 }
 
 pub async fn fetch_available_models(config: &LoongClawConfig) -> CliResult<Vec<String>> {
@@ -152,447 +174,39 @@ fn build_http_client(request_policy: &policy::ProviderRequestPolicy) -> CliResul
         .map_err(|error| format!("build provider http client failed: {error}"))
 }
 
-fn build_completion_request_body(
-    config: &LoongClawConfig,
-    messages: &[Value],
-    model: &str,
-    payload_mode: CompletionPayloadMode,
-) -> Value {
-    let transport_mode = ProviderTransportMode::for_provider(&config.provider);
-    let mut body = serde_json::Map::new();
-    body.insert("model".to_owned(), json!(model));
-    body.insert("messages".to_owned(), Value::Array(messages.to_vec()));
-    body.insert("stream".to_owned(), Value::Bool(false));
-    if payload_mode.temperature_field == TemperatureField::Include {
-        body.insert("temperature".to_owned(), json!(config.provider.temperature));
-    }
-
-    if let Some(limit) = config.provider.max_tokens {
-        match payload_mode.token_field {
-            TokenLimitField::MaxCompletionTokens => {
-                body.insert("max_completion_tokens".to_owned(), json!(limit));
-            }
-            TokenLimitField::MaxTokens => {
-                body.insert("max_tokens".to_owned(), json!(limit));
-            }
-            TokenLimitField::Omit => {}
-        }
-    }
-
-    if let Some(reasoning_effort) = config.provider.reasoning_effort {
-        match payload_mode.reasoning_field {
-            ReasoningField::ReasoningEffort => {
-                body.insert(
-                    "reasoning_effort".to_owned(),
-                    json!(reasoning_effort.as_str()),
-                );
-            }
-            ReasoningField::ReasoningObject => {
-                body.insert(
-                    "reasoning".to_owned(),
-                    json!({
-                        "effort": reasoning_effort.as_str()
-                    }),
-                );
-            }
-            ReasoningField::Omit => {}
-        }
-    }
-
-    if transport_mode == ProviderTransportMode::KimiApi {
-        if let Some(extra_body) = kimi_extra_body(config.provider.reasoning_effort) {
-            body.insert("extra_body".to_owned(), extra_body);
-        }
-    }
-
-    Value::Object(body)
-}
-
-fn build_turn_request_body(
-    config: &LoongClawConfig,
-    messages: &[Value],
-    model: &str,
-    payload_mode: CompletionPayloadMode,
-    include_tool_schema: bool,
-    tool_definitions: &[Value],
-) -> Value {
-    let mut body = build_completion_request_body(config, messages, model, payload_mode);
-    if include_tool_schema && !tool_definitions.is_empty() {
-        if let Some(object) = body.as_object_mut() {
-            object.insert("tools".to_owned(), Value::Array(tool_definitions.to_vec()));
-            object.insert("tool_choice".to_owned(), json!("auto"));
-        }
-    }
-    body
-}
-
-async fn resolve_request_models(
-    config: &LoongClawConfig,
-    headers: &reqwest::header::HeaderMap,
-    request_policy: &policy::ProviderRequestPolicy,
-) -> CliResult<Vec<String>> {
-    if let Some(model) = config.provider.resolved_model() {
-        return Ok(vec![model]);
-    }
-    let available = fetch_available_models_with_policy(config, headers, request_policy).await?;
-    let ordered = rank_model_candidates(&config.provider, &available);
-    if ordered.is_empty() {
-        return Err("provider model-list is empty; set provider.model explicitly".to_owned());
-    }
-    Ok(ordered)
-}
-
-async fn fetch_available_models_with_policy(
-    config: &LoongClawConfig,
-    headers: &reqwest::header::HeaderMap,
-    request_policy: &policy::ProviderRequestPolicy,
-) -> CliResult<Vec<String>> {
-    let endpoint = config.provider.models_endpoint();
-    let client = build_http_client(request_policy)?;
-
-    let mut attempt = 0usize;
-    let mut backoff_ms = request_policy.initial_backoff_ms;
-    loop {
-        attempt += 1;
-        let mut req = client.get(endpoint.clone()).headers(headers.clone());
-        if let Some(auth_header) = config.provider.authorization_header() {
-            req = req.header(reqwest::header::AUTHORIZATION, auth_header);
-        }
-
-        match req.send().await {
-            Ok(response) => {
-                let status = response.status();
-                let response_body = transport::decode_response_body(response)
-                    .await
-                    .map_err(|error| {
-                        format!(
-                            "provider model-list decode failed on attempt {attempt}/{max_attempts}: {error}",
-                            max_attempts = request_policy.max_attempts
-                        )
-                    })?;
-
-                if status.is_success() {
-                    let models = shape::extract_model_ids(&response_body);
-                    if models.is_empty() {
-                        return Err(format!(
-                            "provider model-list returned no models from endpoint `{endpoint}`"
-                        ));
-                    }
-                    return Ok(models);
-                }
-
-                let status_code = status.as_u16();
-                if attempt < request_policy.max_attempts && policy::should_retry_status(status_code)
-                {
-                    sleep(Duration::from_millis(backoff_ms)).await;
-                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
-                    continue;
-                }
-
-                return Err(format!(
-                    "provider model-list returned status {status_code} on attempt {attempt}/{max_attempts}: {response_body}",
-                    max_attempts = request_policy.max_attempts
-                ));
-            }
-            Err(error) => {
-                if attempt < request_policy.max_attempts && policy::should_retry_error(&error) {
-                    sleep(Duration::from_millis(backoff_ms)).await;
-                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
-                    continue;
-                }
-                return Err(format!(
-                    "provider model-list request failed on attempt {attempt}/{max_attempts}: {error}",
-                    max_attempts = request_policy.max_attempts
-                ));
-            }
-        }
-    }
-}
-
-fn rank_model_candidates(provider: &ProviderConfig, available: &[String]) -> Vec<String> {
-    let mut ordered = Vec::new();
-    for raw in &provider.preferred_models {
-        let trimmed = raw.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if let Some(matched) = available.iter().find(|model| *model == trimmed) {
-            push_unique_model(&mut ordered, matched);
-            continue;
-        }
-        if let Some(matched) = available
-            .iter()
-            .find(|model| model.eq_ignore_ascii_case(trimmed))
-        {
-            push_unique_model(&mut ordered, matched);
-        }
-    }
-
-    for model in available {
-        push_unique_model(&mut ordered, model);
-    }
-    ordered
-}
-
-fn push_unique_model(out: &mut Vec<String>, model: &str) {
-    if out.iter().any(|existing| existing == model) {
-        return;
-    }
-    out.push(model.to_owned());
-}
-
 #[derive(Debug)]
 struct ModelRequestError {
     message: String,
     try_next_model: bool,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ProviderTransportMode {
-    OpenAiChatCompletions,
-    KimiApi,
+struct ProviderRequestContext<'a> {
+    endpoint: &'a str,
+    headers: &'a reqwest::header::HeaderMap,
+    request_policy: &'a policy::ProviderRequestPolicy,
+    client: &'a reqwest::Client,
+    auto_model_mode: bool,
 }
 
-impl ProviderTransportMode {
-    fn for_provider(provider: &ProviderConfig) -> Self {
-        match provider.kind {
-            ProviderKind::KimiCoding => Self::KimiApi,
-            _ => Self::OpenAiChatCompletions,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TokenLimitField {
-    MaxCompletionTokens,
-    MaxTokens,
-    Omit,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ReasoningField {
-    ReasoningEffort,
-    ReasoningObject,
-    Omit,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TemperatureField {
-    Include,
-    Omit,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct CompletionPayloadMode {
-    token_field: TokenLimitField,
-    reasoning_field: ReasoningField,
-    temperature_field: TemperatureField,
-}
-
-impl CompletionPayloadMode {
-    fn default_for(provider: &ProviderConfig) -> Self {
-        let token_field = if provider.max_tokens.is_some() {
-            if provider.kind == ProviderKind::Openai {
-                TokenLimitField::MaxCompletionTokens
-            } else {
-                TokenLimitField::MaxTokens
-            }
-        } else {
-            TokenLimitField::Omit
-        };
-
-        let reasoning_field = if provider.reasoning_effort.is_some() {
-            ReasoningField::ReasoningEffort
-        } else {
-            ReasoningField::Omit
-        };
-
-        Self {
-            token_field,
-            reasoning_field,
-            temperature_field: TemperatureField::Include,
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-struct ProviderApiError {
-    code: Option<String>,
-    param: Option<String>,
-    message: Option<String>,
-}
-
-fn parse_provider_api_error(body: &Value) -> ProviderApiError {
-    let error = body.get("error").unwrap_or(body);
-    ProviderApiError {
-        code: error
-            .get("code")
-            .and_then(Value::as_str)
-            .map(str::to_lowercase),
-        param: error
-            .get("param")
-            .and_then(Value::as_str)
-            .map(str::to_lowercase),
-        message: error
-            .get("message")
-            .and_then(Value::as_str)
-            .map(str::to_lowercase),
-    }
-}
-
-fn is_parameter_unsupported(error: &ProviderApiError, parameter: &str) -> bool {
-    let param = parameter.to_lowercase();
-    if error.param.as_deref() == Some(param.as_str()) {
-        return true;
-    }
-
-    let message = error.message.as_deref().unwrap_or_default();
-    if !(message.contains("unknown parameter")
-        || message.contains("unsupported parameter")
-        || message.contains("not supported"))
-    {
-        return false;
-    }
-    message.contains(param.as_str())
-}
-
-fn should_disable_tool_schema_for_error(error: &ProviderApiError) -> bool {
-    if is_parameter_unsupported(error, "tools") || is_parameter_unsupported(error, "tool_choice") {
-        return true;
-    }
-
-    let message = error.message.as_deref().unwrap_or_default();
-    message.contains("tools are not supported")
-        || message.contains("tool use is not supported")
-        || message.contains("function calling is not supported")
-}
-
-fn should_try_next_model_on_error(error: &ProviderApiError) -> bool {
-    if let Some(code) = error.code.as_deref() {
-        if matches!(
-            code,
-            "model_not_found" | "unsupported_model" | "invalid_model" | "not_found_error"
-        ) {
-            return true;
-        }
-    }
-    if error.param.as_deref() == Some("model") {
-        return true;
-    }
-
-    let message = error.message.as_deref().unwrap_or_default();
-    message.contains("only supports")
-        || message.contains("only available")
-        || message.contains("this endpoint")
-        || message.contains("/v1/responses")
-        || message.contains("model does not exist")
-}
-
-fn adapt_payload_mode_for_error(
-    current: CompletionPayloadMode,
-    provider: &ProviderConfig,
-    error: &ProviderApiError,
-) -> Option<CompletionPayloadMode> {
-    if provider.max_tokens.is_some() {
-        if is_parameter_unsupported(error, "max_tokens") {
-            return Some(match current.token_field {
-                TokenLimitField::MaxTokens => CompletionPayloadMode {
-                    token_field: TokenLimitField::MaxCompletionTokens,
-                    ..current
-                },
-                TokenLimitField::MaxCompletionTokens => CompletionPayloadMode {
-                    token_field: TokenLimitField::Omit,
-                    ..current
-                },
-                TokenLimitField::Omit => current,
-            });
-        }
-
-        if is_parameter_unsupported(error, "max_completion_tokens") {
-            return Some(match current.token_field {
-                TokenLimitField::MaxCompletionTokens => CompletionPayloadMode {
-                    token_field: TokenLimitField::MaxTokens,
-                    ..current
-                },
-                TokenLimitField::MaxTokens => CompletionPayloadMode {
-                    token_field: TokenLimitField::Omit,
-                    ..current
-                },
-                TokenLimitField::Omit => current,
-            });
-        }
-    }
-
-    if provider.reasoning_effort.is_some() {
-        if is_parameter_unsupported(error, "reasoning_effort") {
-            return Some(match current.reasoning_field {
-                ReasoningField::ReasoningEffort => CompletionPayloadMode {
-                    reasoning_field: ReasoningField::ReasoningObject,
-                    ..current
-                },
-                ReasoningField::ReasoningObject => CompletionPayloadMode {
-                    reasoning_field: ReasoningField::Omit,
-                    ..current
-                },
-                ReasoningField::Omit => current,
-            });
-        }
-
-        if is_parameter_unsupported(error, "reasoning") {
-            return Some(match current.reasoning_field {
-                ReasoningField::ReasoningObject => CompletionPayloadMode {
-                    reasoning_field: ReasoningField::ReasoningEffort,
-                    ..current
-                },
-                ReasoningField::ReasoningEffort => CompletionPayloadMode {
-                    reasoning_field: ReasoningField::Omit,
-                    ..current
-                },
-                ReasoningField::Omit => current,
-            });
-        }
-    }
-
-    let temperature_rejected = is_parameter_unsupported(error, "temperature")
-        || (error.param.as_deref() == Some("temperature")
-            && error
-                .message
-                .as_deref()
-                .unwrap_or_default()
-                .contains("only the default"));
-    if temperature_rejected {
-        return Some(match current.temperature_field {
-            TemperatureField::Include => CompletionPayloadMode {
-                temperature_field: TemperatureField::Omit,
-                ..current
-            },
-            TemperatureField::Omit => current,
-        });
-    }
-
-    None
-}
-
-#[allow(clippy::too_many_arguments)]
 async fn request_completion_with_model(
     config: &LoongClawConfig,
     messages: &[Value],
     model: &str,
-    auto_model_mode: bool,
-    endpoint: &str,
-    headers: &reqwest::header::HeaderMap,
-    request_policy: &policy::ProviderRequestPolicy,
-    client: &reqwest::Client,
+    request_context: &ProviderRequestContext<'_>,
 ) -> Result<String, ModelRequestError> {
     let mut attempt = 0usize;
-    let mut backoff_ms = request_policy.initial_backoff_ms;
+    let mut backoff_ms = request_context.request_policy.initial_backoff_ms;
     let mut payload_mode = CompletionPayloadMode::default_for(&config.provider);
     let mut tried_payload_modes = vec![payload_mode];
 
     loop {
         attempt += 1;
         let body = build_completion_request_body(config, messages, model, payload_mode);
-        let mut req = client.post(endpoint).headers(headers.clone()).json(&body);
+        let mut req = request_context
+            .client
+            .post(request_context.endpoint)
+            .headers(request_context.headers.clone())
+            .json(&body);
         if let Some(auth_header) = config.provider.authorization_header() {
             req = req.header(reqwest::header::AUTHORIZATION, auth_header);
         }
@@ -605,7 +219,7 @@ async fn request_completion_with_model(
                     .map_err(|error| ModelRequestError {
                         message: format!(
                             "provider response decode failed for model `{model}` on attempt {attempt}/{max_attempts}: {error}",
-                            max_attempts = request_policy.max_attempts
+                            max_attempts = request_context.request_policy.max_attempts
                         ),
                         try_next_model: false,
                     })?;
@@ -615,7 +229,7 @@ async fn request_completion_with_model(
                         ModelRequestError {
                             message: format!(
                                 "provider response missing choices[0].message.content for model `{model}` on attempt {attempt}/{max_attempts}: {response_body}",
-                                max_attempts = request_policy.max_attempts
+                                max_attempts = request_context.request_policy.max_attempts
                             ),
                             try_next_model: false,
                         }
@@ -635,14 +249,18 @@ async fn request_completion_with_model(
                 }
 
                 let status_code = status.as_u16();
-                if attempt < request_policy.max_attempts && policy::should_retry_status(status_code)
+                if attempt < request_context.request_policy.max_attempts
+                    && policy::should_retry_status(status_code)
                 {
                     sleep(Duration::from_millis(backoff_ms)).await;
-                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
+                    backoff_ms = policy::next_backoff_ms(
+                        backoff_ms,
+                        request_context.request_policy.max_backoff_ms,
+                    );
                     continue;
                 }
 
-                if auto_model_mode && should_try_next_model_on_error(&api_error) {
+                if request_context.auto_model_mode && should_try_next_model_on_error(&api_error) {
                     return Err(ModelRequestError {
                         message: format!(
                             "model `{model}` rejected by provider endpoint; trying next candidate. status {status_code}: {response_body}"
@@ -654,21 +272,26 @@ async fn request_completion_with_model(
                 return Err(ModelRequestError {
                     message: format!(
                         "provider returned status {status_code} for model `{model}` on attempt {attempt}/{max_attempts}: {response_body}",
-                        max_attempts = request_policy.max_attempts
+                        max_attempts = request_context.request_policy.max_attempts
                     ),
                     try_next_model: false,
                 });
             }
             Err(error) => {
-                if attempt < request_policy.max_attempts && policy::should_retry_error(&error) {
+                if attempt < request_context.request_policy.max_attempts
+                    && policy::should_retry_error(&error)
+                {
                     sleep(Duration::from_millis(backoff_ms)).await;
-                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
+                    backoff_ms = policy::next_backoff_ms(
+                        backoff_ms,
+                        request_context.request_policy.max_backoff_ms,
+                    );
                     continue;
                 }
                 return Err(ModelRequestError {
                     message: format!(
                         "provider request failed for model `{model}` on attempt {attempt}/{max_attempts}: {error}",
-                        max_attempts = request_policy.max_attempts
+                        max_attempts = request_context.request_policy.max_attempts
                     ),
                     try_next_model: false,
                 });
@@ -677,19 +300,14 @@ async fn request_completion_with_model(
     }
 }
 
-#[allow(clippy::too_many_arguments)]
 async fn request_turn_with_model(
     config: &LoongClawConfig,
     messages: &[Value],
     model: &str,
-    auto_model_mode: bool,
-    endpoint: &str,
-    headers: &reqwest::header::HeaderMap,
-    request_policy: &policy::ProviderRequestPolicy,
-    client: &reqwest::Client,
+    request_context: &ProviderRequestContext<'_>,
 ) -> Result<crate::conversation::turn_engine::ProviderTurn, ModelRequestError> {
     let mut attempt = 0usize;
-    let mut backoff_ms = request_policy.initial_backoff_ms;
+    let mut backoff_ms = request_context.request_policy.initial_backoff_ms;
     let mut payload_mode = CompletionPayloadMode::default_for(&config.provider);
     let mut tried_payload_modes = vec![payload_mode];
     let tool_definitions = super::tools::provider_tool_definitions();
@@ -705,7 +323,11 @@ async fn request_turn_with_model(
             include_tool_schema,
             &tool_definitions,
         );
-        let mut req = client.post(endpoint).headers(headers.clone()).json(&body);
+        let mut req = request_context
+            .client
+            .post(request_context.endpoint)
+            .headers(request_context.headers.clone())
+            .json(&body);
         if let Some(auth_header) = config.provider.authorization_header() {
             req = req.header(reqwest::header::AUTHORIZATION, auth_header);
         }
@@ -718,7 +340,7 @@ async fn request_turn_with_model(
                     .map_err(|error| ModelRequestError {
                         message: format!(
                             "provider response decode failed for model `{model}` on attempt {attempt}/{max_attempts}: {error}",
-                            max_attempts = request_policy.max_attempts
+                            max_attempts = request_context.request_policy.max_attempts
                         ),
                         try_next_model: false,
                     })?;
@@ -728,7 +350,7 @@ async fn request_turn_with_model(
                         ModelRequestError {
                             message: format!(
                                 "provider response missing choices[0].message for model `{model}` on attempt {attempt}/{max_attempts}: {response_body}",
-                                max_attempts = request_policy.max_attempts
+                                max_attempts = request_context.request_policy.max_attempts
                             ),
                             try_next_model: false,
                         }
@@ -752,14 +374,18 @@ async fn request_turn_with_model(
                 }
 
                 let status_code = status.as_u16();
-                if attempt < request_policy.max_attempts && policy::should_retry_status(status_code)
+                if attempt < request_context.request_policy.max_attempts
+                    && policy::should_retry_status(status_code)
                 {
                     sleep(Duration::from_millis(backoff_ms)).await;
-                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
+                    backoff_ms = policy::next_backoff_ms(
+                        backoff_ms,
+                        request_context.request_policy.max_backoff_ms,
+                    );
                     continue;
                 }
 
-                if auto_model_mode && should_try_next_model_on_error(&api_error) {
+                if request_context.auto_model_mode && should_try_next_model_on_error(&api_error) {
                     return Err(ModelRequestError {
                         message: format!(
                             "model `{model}` rejected by provider endpoint; trying next candidate. status {status_code}: {response_body}"
@@ -771,21 +397,26 @@ async fn request_turn_with_model(
                 return Err(ModelRequestError {
                     message: format!(
                         "provider returned status {status_code} for model `{model}` on attempt {attempt}/{max_attempts}: {response_body}",
-                        max_attempts = request_policy.max_attempts
+                        max_attempts = request_context.request_policy.max_attempts
                     ),
                     try_next_model: false,
                 });
             }
             Err(error) => {
-                if attempt < request_policy.max_attempts && policy::should_retry_error(&error) {
+                if attempt < request_context.request_policy.max_attempts
+                    && policy::should_retry_error(&error)
+                {
                     sleep(Duration::from_millis(backoff_ms)).await;
-                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
+                    backoff_ms = policy::next_backoff_ms(
+                        backoff_ms,
+                        request_context.request_policy.max_backoff_ms,
+                    );
                     continue;
                 }
                 return Err(ModelRequestError {
                     message: format!(
                         "provider request failed for model `{model}` on attempt {attempt}/{max_attempts}: {error}",
-                        max_attempts = request_policy.max_attempts
+                        max_attempts = request_context.request_policy.max_attempts
                     ),
                     try_next_model: false,
                 });
@@ -794,92 +425,14 @@ async fn request_turn_with_model(
     }
 }
 
-fn validate_provider_feature_gate(config: &LoongClawConfig) -> CliResult<()> {
-    match config.provider.kind {
-        ProviderKind::Volcengine => {
-            if !cfg!(feature = "provider-volcengine") {
-                return Err(
-                    "volcengine provider is disabled (enable feature `provider-volcengine`)"
-                        .to_owned(),
-                );
-            }
-        }
-        _ => {
-            if !cfg!(feature = "provider-openai") {
-                return Err(
-                    "openai-compatible provider family is disabled (enable feature `provider-openai`)"
-                        .to_owned(),
-                );
-            }
-        }
-    }
-    Ok(())
-}
-
-fn validate_provider_configuration(config: &LoongClawConfig) -> CliResult<()> {
-    if config.provider.kind == ProviderKind::Kimi
-        && provider_uses_kimi_coding_endpoint(&config.provider)
-    {
-        return Err(
-            "kimi provider cannot target Kimi Coding endpoints; use `kind = \"kimi_coding\"`"
-                .to_owned(),
-        );
-    }
-
-    if config.provider.kind == ProviderKind::KimiCoding {
-        if let Some(user_agent) = config.provider.header_value("user-agent") {
-            if !is_kimi_cli_user_agent(user_agent) {
-                return Err(format!(
-                    "kimi_coding provider requires a `User-Agent` header starting with `KimiCLI/`; got `{user_agent}`"
-                ));
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn provider_uses_kimi_coding_endpoint(provider: &ProviderConfig) -> bool {
-    is_kimi_coding_endpoint(provider.endpoint().as_str())
-        || provider
-            .endpoint
-            .as_deref()
-            .is_some_and(is_kimi_coding_endpoint)
-}
-
-fn is_kimi_coding_endpoint(endpoint: &str) -> bool {
-    endpoint
-        .trim()
-        .to_ascii_lowercase()
-        .contains("://api.kimi.com/coding/")
-}
-
-fn is_kimi_cli_user_agent(user_agent: &str) -> bool {
-    user_agent.trim().starts_with("KimiCLI/")
-}
-
-fn kimi_extra_body(reasoning_effort: Option<ReasoningEffort>) -> Option<Value> {
-    let reasoning_effort = reasoning_effort?;
-    let thinking_type = match reasoning_effort {
-        ReasoningEffort::None => "disabled",
-        ReasoningEffort::Minimal
-        | ReasoningEffort::Low
-        | ReasoningEffort::Medium
-        | ReasoningEffort::High
-        | ReasoningEffort::Xhigh => "enabled",
-    };
-    Some(json!({
-        "thinking": {
-            "type": thinking_type
-        }
-    }))
-}
-
 #[cfg(test)]
 mod tests {
+    use super::model_selection::rank_model_candidates;
+    use super::payload_adaptation::{ReasoningField, TemperatureField, TokenLimitField};
     use super::*;
     use crate::config::{
-        FeishuChannelConfig, MemoryConfig, ProviderConfig, ReasoningEffort, ToolConfig,
+        FeishuChannelConfig, MemoryConfig, ProviderConfig, ProviderKind, ReasoningEffort,
+        ToolConfig,
     };
     use serde_json::json;
