+++ crates/app/src/provider/model_selection.rs
@@ -1,0 +1,126 @@
+use std::time::Duration;
+
+use tokio::time::sleep;
+
+use crate::{config::ProviderConfig, CliResult};
+
+use super::{build_http_client, policy, shape, transport};
+use crate::config::LoongClawConfig;
+
+pub(super) async fn resolve_request_models(
+    config: &LoongClawConfig,
+    headers: &reqwest::header::HeaderMap,
+    request_policy: &policy::ProviderRequestPolicy,
+) -> CliResult<Vec<String>> {
+    if let Some(model) = config.provider.resolved_model() {
+        return Ok(vec![model]);
+    }
+    let available = fetch_available_models_with_policy(config, headers, request_policy).await?;
+    let ordered = rank_model_candidates(&config.provider, &available);
+    if ordered.is_empty() {
+        return Err("provider model-list is empty; set provider.model explicitly".to_owned());
+    }
+    Ok(ordered)
+}
+
+pub(super) async fn fetch_available_models_with_policy(
+    config: &LoongClawConfig,
+    headers: &reqwest::header::HeaderMap,
+    request_policy: &policy::ProviderRequestPolicy,
+) -> CliResult<Vec<String>> {
+    let endpoint = config.provider.models_endpoint();
+    let client = build_http_client(request_policy)?;
+
+    let mut attempt = 0usize;
+    let mut backoff_ms = request_policy.initial_backoff_ms;
+    loop {
+        attempt += 1;
+        let mut req = client.get(endpoint.clone()).headers(headers.clone());
+        if let Some(auth_header) = config.provider.authorization_header() {
+            req = req.header(reqwest::header::AUTHORIZATION, auth_header);
+        }
+
+        match req.send().await {
+            Ok(response) => {
+                let status = response.status();
+                let response_body = transport::decode_response_body(response)
+                    .await
+                    .map_err(|error| {
+                        format!(
+                            "provider model-list decode failed on attempt {attempt}/{max_attempts}: {error}",
+                            max_attempts = request_policy.max_attempts
+                        )
+                    })?;
+
+                if status.is_success() {
+                    let models = shape::extract_model_ids(&response_body);
+                    if models.is_empty() {
+                        return Err(format!(
+                            "provider model-list returned no models from endpoint `{endpoint}`"
+                        ));
+                    }
+                    return Ok(models);
+                }
+
+                let status_code = status.as_u16();
+                if attempt < request_policy.max_attempts && policy::should_retry_status(status_code)
+                {
+                    sleep(Duration::from_millis(backoff_ms)).await;
+                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
+                    continue;
+                }
+
+                return Err(format!(
+                    "provider model-list returned status {status_code} on attempt {attempt}/{max_attempts}: {response_body}",
+                    max_attempts = request_policy.max_attempts
+                ));
+            }
+            Err(error) => {
+                if attempt < request_policy.max_attempts && policy::should_retry_error(&error) {
+                    sleep(Duration::from_millis(backoff_ms)).await;
+                    backoff_ms = policy::next_backoff_ms(backoff_ms, request_policy.max_backoff_ms);
+                    continue;
+                }
+                return Err(format!(
+                    "provider model-list request failed on attempt {attempt}/{max_attempts}: {error}",
+                    max_attempts = request_policy.max_attempts
+                ));
+            }
+        }
+    }
+}
+
+pub(super) fn rank_model_candidates(
+    provider: &ProviderConfig,
+    available: &[String],
+) -> Vec<String> {
+    let mut ordered = Vec::new();
+    for raw in &provider.preferred_models {
+        let trimmed = raw.trim();
+        if trimmed.is_empty() {
+            continue;
+        }
+        if let Some(matched) = available.iter().find(|model| *model == trimmed) {
+            push_unique_model(&mut ordered, matched);
+            continue;
+        }
+        if let Some(matched) = available
+            .iter()
+            .find(|model| model.eq_ignore_ascii_case(trimmed))
+        {
+            push_unique_model(&mut ordered, matched);
+        }
+    }
+
+    for model in available {
+        push_unique_model(&mut ordered, model);
+    }
+    ordered
+}
+
+fn push_unique_model(out: &mut Vec<String>, model: &str) {
+    if out.iter().any(|existing| existing == model) {
+        return;
+    }
+    out.push(model.to_owned());
+}
