+++ crates/spec/src/spec_execution/tool_search.rs
@@ -1,0 +1,358 @@
+use std::collections::BTreeMap;
+
+use kernel::{IntegrationCatalog, PluginBridgeKind, PluginScanReport, PluginTranslationReport};
+use serde_json::Value;
+
+use super::descriptor_bridge_kind;
+use crate::spec_runtime::{detect_provider_bridge_kind, ToolSearchEntry, ToolSearchResult};
+
+pub(super) fn execute_tool_search(
+    integration_catalog: &IntegrationCatalog,
+    plugin_scan_reports: &[PluginScanReport],
+    plugin_translation_reports: &[PluginTranslationReport],
+    query: &str,
+    limit: usize,
+    include_deferred: bool,
+    include_examples: bool,
+) -> Vec<ToolSearchResult> {
+    let mut entries: BTreeMap<String, ToolSearchEntry> = BTreeMap::new();
+    let mut translation_by_key: BTreeMap<
+        (String, String),
+        (PluginBridgeKind, String, String, String),
+    > = BTreeMap::new();
+
+    for report in plugin_translation_reports {
+        for entry in &report.entries {
+            translation_by_key.insert(
+                (entry.source_path.clone(), entry.plugin_id.clone()),
+                (
+                    entry.runtime.bridge_kind,
+                    entry.runtime.adapter_family.clone(),
+                    entry.runtime.entrypoint_hint.clone(),
+                    entry.runtime.source_language.clone(),
+                ),
+            );
+        }
+    }
+
+    for provider in integration_catalog.providers() {
+        let channel_endpoint = integration_catalog
+            .channels_for_provider(&provider.provider_id)
+            .into_iter()
+            .find(|channel| channel.enabled)
+            .map(|channel| channel.endpoint)
+            .unwrap_or_default();
+        let bridge_kind = detect_provider_bridge_kind(&provider, &channel_endpoint);
+        let tool_id = format!("{}::{}", provider.provider_id, provider.connector_name);
+        let summary = provider.metadata.get("summary").cloned();
+        let tags = metadata_tags(&provider.metadata);
+        let input_examples = metadata_examples(&provider.metadata, "input_examples_json");
+        let output_examples = metadata_examples(&provider.metadata, "output_examples_json");
+        let deferred = metadata_bool(&provider.metadata, "defer_loading").unwrap_or(false);
+        let mut adapter_family = provider.metadata.get("adapter_family").cloned();
+        let mut entrypoint_hint = provider
+            .metadata
+            .get("entrypoint")
+            .or_else(|| provider.metadata.get("entrypoint_hint"))
+            .cloned();
+        let mut source_language = provider.metadata.get("source_language").cloned();
+        let mut resolved_bridge_kind = bridge_kind;
+
+        if let (Some(source_path), Some(plugin_id)) = (
+            provider.metadata.get("plugin_source_path"),
+            provider.metadata.get("plugin_id"),
+        ) {
+            if let Some((bridge, adapter, entrypoint, language)) =
+                translation_by_key.get(&(source_path.clone(), plugin_id.clone()))
+            {
+                resolved_bridge_kind = *bridge;
+                adapter_family = Some(adapter.clone());
+                entrypoint_hint = Some(entrypoint.clone());
+                source_language = Some(language.clone());
+            }
+        }
+
+        entries.insert(
+            tool_id.clone(),
+            ToolSearchEntry {
+                tool_id,
+                plugin_id: provider.metadata.get("plugin_id").cloned(),
+                connector_name: provider.connector_name.clone(),
+                provider_id: provider.provider_id.clone(),
+                source_path: provider.metadata.get("plugin_source_path").cloned(),
+                bridge_kind: resolved_bridge_kind,
+                adapter_family,
+                entrypoint_hint,
+                source_language,
+                summary,
+                tags,
+                input_examples,
+                output_examples,
+                deferred,
+                loaded: true,
+            },
+        );
+    }
+
+    for report in plugin_scan_reports {
+        for descriptor in &report.descriptors {
+            let manifest = &descriptor.manifest;
+            let tool_id = format!("{}::{}", manifest.provider_id, manifest.connector_name);
+            let translation =
+                translation_by_key.get(&(descriptor.path.clone(), manifest.plugin_id.clone()));
+            let bridge_kind = translation
+                .map(|(bridge, _, _, _)| *bridge)
+                .unwrap_or_else(|| descriptor_bridge_kind(descriptor));
+            let adapter_family = translation.map(|(_, adapter, _, _)| adapter.clone());
+            let entrypoint_hint = translation.map(|(_, _, entrypoint, _)| entrypoint.clone());
+            let source_language = translation.map(|(_, _, _, language)| language.clone());
+
+            let entry = entries
+                .entry(tool_id.clone())
+                .or_insert_with(|| ToolSearchEntry {
+                    tool_id: tool_id.clone(),
+                    plugin_id: Some(manifest.plugin_id.clone()),
+                    connector_name: manifest.connector_name.clone(),
+                    provider_id: manifest.provider_id.clone(),
+                    source_path: Some(descriptor.path.clone()),
+                    bridge_kind,
+                    adapter_family: adapter_family.clone(),
+                    entrypoint_hint: entrypoint_hint.clone(),
+                    source_language: source_language.clone(),
+                    summary: manifest.summary.clone(),
+                    tags: manifest.tags.clone(),
+                    input_examples: manifest.input_examples.clone(),
+                    output_examples: manifest.output_examples.clone(),
+                    deferred: manifest.defer_loading,
+                    loaded: false,
+                });
+
+            if entry.plugin_id.is_none() {
+                entry.plugin_id = Some(manifest.plugin_id.clone());
+            }
+            if entry.source_path.is_none() {
+                entry.source_path = Some(descriptor.path.clone());
+            }
+            if entry.summary.is_none() {
+                entry.summary = manifest.summary.clone();
+            }
+            if entry.adapter_family.is_none() {
+                entry.adapter_family = adapter_family.clone();
+            }
+            if entry.entrypoint_hint.is_none() {
+                entry.entrypoint_hint = entrypoint_hint.clone();
+            }
+            if entry.source_language.is_none() {
+                entry.source_language = source_language.clone();
+            }
+            if entry.input_examples.is_empty() {
+                entry.input_examples = manifest.input_examples.clone();
+            }
+            if entry.output_examples.is_empty() {
+                entry.output_examples = manifest.output_examples.clone();
+            }
+            for tag in &manifest.tags {
+                if !entry.tags.iter().any(|existing| existing == tag) {
+                    entry.tags.push(tag.clone());
+                }
+            }
+            if !entry.loaded {
+                entry.deferred = manifest.defer_loading;
+                entry.bridge_kind = bridge_kind;
+            }
+        }
+    }
+
+    let query_normalized = query.trim().to_ascii_lowercase();
+    let tokens: Vec<String> = query_normalized
+        .split(|ch: char| !ch.is_ascii_alphanumeric() && ch != '_' && ch != '-')
+        .map(str::trim)
+        .filter(|token| !token.is_empty())
+        .map(str::to_owned)
+        .collect();
+
+    let mut ranked: Vec<(u32, ToolSearchEntry)> = entries
+        .into_values()
+        .filter(|entry| include_deferred || !entry.deferred || entry.loaded)
+        .filter_map(|entry| {
+            let score = tool_search_score(&entry, &query_normalized, &tokens);
+            if query_normalized.is_empty() || score > 0 {
+                Some((score, entry))
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    ranked.sort_by(|(left_score, left), (right_score, right)| {
+        right_score
+            .cmp(left_score)
+            .then_with(|| right.loaded.cmp(&left.loaded))
+            .then_with(|| left.tool_id.cmp(&right.tool_id))
+    });
+
+    let capped_limit = limit.clamp(1, 50);
+    ranked
+        .into_iter()
+        .take(capped_limit)
+        .map(|(score, entry)| ToolSearchResult {
+            tool_id: entry.tool_id,
+            plugin_id: entry.plugin_id,
+            connector_name: entry.connector_name,
+            provider_id: entry.provider_id,
+            source_path: entry.source_path,
+            bridge_kind: entry.bridge_kind.as_str().to_owned(),
+            adapter_family: entry.adapter_family,
+            entrypoint_hint: entry.entrypoint_hint,
+            source_language: entry.source_language,
+            score,
+            deferred: entry.deferred,
+            loaded: entry.loaded,
+            summary: entry.summary,
+            tags: entry.tags,
+            input_examples: if include_examples {
+                entry.input_examples
+            } else {
+                Vec::new()
+            },
+            output_examples: if include_examples {
+                entry.output_examples
+            } else {
+                Vec::new()
+            },
+        })
+        .collect()
+}
+
+fn metadata_tags(metadata: &BTreeMap<String, String>) -> Vec<String> {
+    if let Some(raw_json) = metadata.get("tags_json") {
+        if let Ok(values) = serde_json::from_str::<Vec<String>>(raw_json) {
+            return values;
+        }
+    }
+
+    metadata
+        .get("tags")
+        .map(|raw| {
+            raw.split([',', ';'])
+                .map(str::trim)
+                .filter(|value| !value.is_empty())
+                .map(str::to_owned)
+                .collect::<Vec<_>>()
+        })
+        .unwrap_or_default()
+}
+
+fn metadata_examples(metadata: &BTreeMap<String, String>, key: &str) -> Vec<Value> {
+    metadata
+        .get(key)
+        .and_then(|raw| serde_json::from_str::<Vec<Value>>(raw).ok())
+        .unwrap_or_default()
+}
+
+fn metadata_bool(metadata: &BTreeMap<String, String>, key: &str) -> Option<bool> {
+    metadata
+        .get(key)
+        .and_then(|raw| match raw.trim().to_ascii_lowercase().as_str() {
+            "true" | "1" | "yes" | "y" | "on" => Some(true),
+            "false" | "0" | "no" | "n" | "off" => Some(false),
+            _ => None,
+        })
+}
+
+fn tool_search_score(entry: &ToolSearchEntry, query: &str, tokens: &[String]) -> u32 {
+    if query.is_empty() {
+        return if entry.loaded { 10 } else { 5 };
+    }
+
+    let connector = entry.connector_name.to_ascii_lowercase();
+    let provider = entry.provider_id.to_ascii_lowercase();
+    let tool_id = entry.tool_id.to_ascii_lowercase();
+    let summary = entry
+        .summary
+        .as_deref()
+        .unwrap_or_default()
+        .to_ascii_lowercase();
+    let source_path = entry
+        .source_path
+        .as_deref()
+        .unwrap_or_default()
+        .to_ascii_lowercase();
+    let adapter_family = entry
+        .adapter_family
+        .as_deref()
+        .unwrap_or_default()
+        .to_ascii_lowercase();
+    let entrypoint_hint = entry
+        .entrypoint_hint
+        .as_deref()
+        .unwrap_or_default()
+        .to_ascii_lowercase();
+    let source_language = entry
+        .source_language
+        .as_deref()
+        .unwrap_or_default()
+        .to_ascii_lowercase();
+    let tags: Vec<String> = entry
+        .tags
+        .iter()
+        .map(|tag| tag.to_ascii_lowercase())
+        .collect();
+
+    let mut score = 0_u32;
+    if connector == query {
+        score = score.saturating_add(150);
+    } else if connector.contains(query) {
+        score = score.saturating_add(110);
+    }
+    if provider == query {
+        score = score.saturating_add(120);
+    } else if provider.contains(query) {
+        score = score.saturating_add(80);
+    }
+    if tool_id.contains(query) {
+        score = score.saturating_add(60);
+    }
+    if summary.contains(query) {
+        score = score.saturating_add(55);
+    }
+    if source_path.contains(query) {
+        score = score.saturating_add(35);
+    }
+    if adapter_family.contains(query) {
+        score = score.saturating_add(18);
+    }
+    if entrypoint_hint.contains(query) {
+        score = score.saturating_add(12);
+    }
+    if source_language.contains(query) {
+        score = score.saturating_add(10);
+    }
+    if tags.iter().any(|tag| tag == query) {
+        score = score.saturating_add(45);
+    } else if tags.iter().any(|tag| tag.contains(query)) {
+        score = score.saturating_add(25);
+    }
+
+    let haystack = format!(
+        "{} {} {} {} {} {} {} {}",
+        connector,
+        provider,
+        tool_id,
+        summary,
+        adapter_family,
+        entrypoint_hint,
+        source_language,
+        tags.join(" ")
+    );
+    for token in tokens {
+        if haystack.contains(token) {
+            score = score.saturating_add(8);
+        }
+    }
+
+    if entry.loaded {
+        score = score.saturating_add(4);
+    }
+    score
+}
